+++ test/DryIoc.IssuesTests/TestTools.cs
@@ -1,0 +1,7 @@
+namespace DryIoc.IssuesTests
+{
+    public interface ITest
+    {
+        int Run();
+    }
+}